Allow for a null value
This small change allows a user to optionally let the colorPicker be used safely on an input field that defaults to null so long as it is invoked with option allowNull set to true. May require a bit more work to check that it affects nothing else but it nominally works in my early testing.